+++ packages/client/src/Pages/Topic/components/Comment/Comment.tsx
@@ -2,7 +2,8 @@ import { useMemo } from 'react';
 import classNames from 'classnames';
 import s from './Comment.module.scss';
 import { IComment } from '@models/types/topics';
-import { CURRENT_USER_ID, USERS } from '../../lib/constants';
+import { CURRENT_USER_ID } from '../../lib/constants';
+import { USERS } from '../../lib/mocks';
 import { SmileMenu } from '../SmileMenu/SmileMenu';
 import { Spacer } from '@/components';
